[fix] tag码识别id范围

+++ src/ai_camera.h
@@ -68,7 +68,7 @@ public:
     String get_qrcode_content(void);
     uint8_t get_identify_num(AI_CAMERA_REGISTER_t features, uint8_t total=0);
     uint8_t get_face_attributes(int &is_male, int &is_mouth_open, int &is_smail, int &is_glasses, uint8_t index=0);
-    uint8_t get_identify_id(AI_CAMERA_REGISTER_t features, uint8_t index=0);
+    uint16_t get_identify_id(AI_CAMERA_REGISTER_t features, uint8_t index=0);
     int16_t get_identify_rotation(AI_CAMERA_REGISTER_t features, uint8_t index=0);
     void get_identify_position(AI_CAMERA_REGISTER_t features, int position[4], uint8_t index=0);
     void get_identify_position(AI_CAMERA_REGISTER_t features, int &x, int &y, int &w, int &h, uint8_t index=0);
